tauclubs/requirements.txt
django
djangorestframework
markdown
django.core.validators